+++ CHANGES.md
@@ -24,6 +24,8 @@
   - @miosakuma
 - [ADD] VideoCodec に H265 を追加する
   - @enm10k
+- [ADD] WebRTCConfiguration に degradationPreference を追加する
+  - @enm10k
 - [UPDATE] 解像度に qHD (960x540) を追加する
   - @enm10k
 - [UPDATE] CocoaPods を v1.14.2 に更新する